Changed census module decorators to overrides
