--- decidim-dipgra_census_authorization/app/decorators/forms/update_organization_form_decorator.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-Decidim::System::UpdateOrganizationForm.class_eval do
-  attribute :ine_code, Integer
-  attribute :municipality_code, Integer
-end
