+++ decidim-dipgra_census_authorization/app/forms/concerns/decidim/system/update_organization_form_override.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Decidim
+  module System
+    module UpdateOrganizationFormOverride
+      extend ActiveSupport::Concern
+
+      included do
+        attribute :ine_code, Integer
+        attribute :municipality_code, Integer
+      end
+    end
+  end
+end
