+++ decidim-dipgra_census_authorization/app/overrides/decidim/system/organizations/add_dipgra_census_settings.rb
@@ -2,5 +2,5 @@
 
 Deface::Override.new(virtual_path: "decidim/system/organizations/_advanced_settings",
                      name: "add_dipgra_census_settings",
-                     insert_before: "erb[loud]:contains('file_upload_settings')",
-                     text: "<%= render partial: 'dipgra_census_settings', locals: { f: f } %>")
+                     insert_before: "erb[silent]:contains('each do |partial|')",
+                     text: "<%= render partial: 'dipgra_census_settings', locals: { f: } %>")
