+++ decidim-dipgra_census_authorization/app/views/decidim/system/organizations/_dipgra_census_settings.html.erb
@@ -1,5 +1,5 @@
 <div class="fieldset">
-  <h4><%= t("decidim.system.models.organization.fields.dipgra_census_settings") %></h4>
+  <legend><label><%= t("decidim.system.models.organization.fields.dipgra_census_settings") %></label></legend>
   <div class="field">
     <%= f.text_field :ine_code %>
   </div>
